+++ src/grocker/__main__.py
@@ -1,9 +1,5 @@
 #! /usr/bin/env python
-# -*- coding: utf-8 -*-
 # Copyright (c) Polyconseil SAS. All rights reserved.
-
-from __future__ import absolute_import
-from __future__ import unicode_literals
 
 import logging