+++ CHANGELOG.rst
@@ -1,8 +1,8 @@
 ChangeLog
 =========
 
-6.10 (unreleased)
------------------
+7.0 (2021-08-13)
+----------------
 
 - Drop Python 3.5 support since it is not supported anymore (**breaking change**).
 - Add Python 3.9 support (for Grocker command).